feat: Palette Picker, Image Source fixes, and Ratio refinements
- Added PalettePickerPage allow editing and creation of custom palettes
- Refactored Ratio settings to support decimal values
- Implemented FromSource interface to reliably track image sources for palettes
- Fixed ImageSampler to correctly capture redirected URLs
- Fixed 'Add' button in history to behave contextually
- Fixed Image Reference overlay to support in-memory images

+++ lib/pages/settings_page.dart
@@ -31,10 +31,12 @@ class _SettingsPageState extends State<SettingsPage> {
 
   Future<void> _loadSettings() async {
     final prefs = await SharedPreferences.getInstance();
+
     setState(() {
       _widthController.text = (prefs.getInt("image_width") ?? 800).toString();
       _heightController.text = (prefs.getInt("image_height") ?? 600).toString();
-      _sizeRatioController.text = (prefs.getInt("size_ratio") ?? 12).toString();
+      _sizeRatioController.text =
+          (prefs.getDouble("size_ratio") ?? 12.0).toString();
       _addTriangleGradients = prefs.getBool("add_triangle_gradients") ?? true;
       _annotateWithDimensions =
           prefs.getBool("annotate_with_dimensions") ?? false;
@@ -47,8 +49,8 @@ class _SettingsPageState extends State<SettingsPage> {
         "image_width", int.tryParse(_widthController.text) ?? 800);
     await prefs.setInt(
         "image_height", int.tryParse(_heightController.text) ?? 600);
-    await prefs.setInt(
-        "size_ratio", int.tryParse(_sizeRatioController.text) ?? 12);
+    await prefs.setDouble(
+        "size_ratio", double.tryParse(_sizeRatioController.text) ?? 12.0);
     await prefs.setBool("add_triangle_gradients", _addTriangleGradients);
     await prefs.setBool("annotate_with_dimensions", _annotateWithDimensions);
     if (mounted) {
